% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visibility_dop.R
\name{calculate_dop}
\alias{calculate_dop}
\title{Calculate expected DOP for a location}
\usage{
calculate_dop(lat_deg, lon_deg, alt, y, m, d, h, mi, sec,
  Obstruction = 0, almanac_file = NULL)
}
\arguments{
\item{lat_deg}{target position latitude in degrees N}

\item{lon_deg}{target position longitude in degrees E}

\item{alt}{altitude (above sealevel) in meters}

\item{y}{year}

\item{m}{month}

\item{d}{day of month}

\item{h}{hour}

\item{mi}{minute}

\item{sec}{second}

\item{Obstruction}{obstruction angle to horizon in degrees. this is currently assumed to be uniform around the target location}

\item{almanac_file}{path to an appropriate almanace file. If NULL (default)}
}
\value{
a named list
}
\description{
Calculate expected DOP for a location
}
\examples{
 # use local almanac file
 calculate_dop(0,90,0,2018,10,4,0,0,0,
  almanac_file = system.file('extdata', 'current.al3', package = 'dopcalc'))
 # get almanac file from USGS online
 calculate_dop(0,90,0,2018,10,4,0,0,0)
}
