% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_almanac.R
\name{get_almanac_url}
\alias{get_almanac_url}
\title{Attempt to find appropriate almanac file for a given date}
\usage{
get_almanac_url(year, month, day, gps_week)
}
\arguments{
\item{year}{numeric year}

\item{month}{numeric month}

\item{day}{numeric day of month}

\item{gps_week}{numeric gps_week}
}
\value{
a string
}
\description{
Attempt to find appropriate almanac file for a given date
}
