% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymdhms2gps.R
\name{ymdhms2gps}
\alias{ymdhms2gps}
\title{Convert a ymdhms date to the date format used by the GPS almanac}
\usage{
ymdhms2gps(year, month, mday, hour, minute, second)
}
\arguments{
\item{year}{integer year as YYYY}

\item{month}{integer month}

\item{mday}{integer day of month}

\item{hour}{integer hour 0-24}

\item{minute}{integer minute}

\item{second}{integer second}
}
\value{
a list with the elements gps_week, the GPS week number,
and sec_of_week, the GPS time of applicability.
}
\description{
Inputs are currently assumed to be in UTC
}
\details{
The GPS week number or almanac reference number is the
 number of weeks since 1980-01-06 modulo 1024. The GPS time of
 applicability is the number of seconds since the beginning of
 the almanac reference week.  See ICD-GPS-240A for further details.
}
\references{
Global Positioning Systems Directorate (2018). Systems Engineering and Integration Interface Specification IS-GPS-200J NAVSTAR GPS Space Segment/Navigation User Segment Interfaces. \href{https://www.gps.gov/technical/icwg/IS-GPS-200J.pdf}{https://www.gps.gov/technical/icwg/IS-GPS-200J.pdf}
}
\author{
Benjamin W. Remondi (original Fortran code), Philipp Boersch-Supan (R translation)
}
